feature(student): adds functionality for student management.
This commit adds the functionality to manage students. It adds a
hibernate model which will be used in storing student data to the
db.

It also adds sample functionality for reference on how to query
the db for students and add students to the db.

It adds a migration file to be used by flyway db to create tables.

It also adds a pom.xml file which contains dependencies for the
program.